+++ src/terms/match/lib/fastPass.js
@@ -5,7 +5,8 @@ const fastPass = (ts, regs) => {
   for(let i = 0; i < regs.length; i++) {
     let reg = regs[i];
     let found = false;
-    if (reg.optional === true || reg.negative === true) {
+    //we can't cheat on these fancy rules:
+    if (reg.optional === true || reg.negative === true || reg.minMax !== undefined) {
       continue;
     }
     //look-for missing term-matches