+++ changelog.md
@@ -10,6 +10,7 @@ here, 'Major' is considered an api change, while 'Minor' is considered a perform
 - new `.lessThan()`, `.greaterThan()`, `.equalTo()` methods
 - new prefix/suffix/infix matches with `_ffix` syntax
 - `tag()` supports a sequence of tags for a sequence of terms
+- .match 'range' queries now use a real match - `#Adverb{2,4}`
 
 ### v.9  &nbsp;
 * 9.0.0